+++ test/LionWeb.Core.Test/Utilities/Comparer/ComparerNodeTests.cs
@@ -20,6 +20,7 @@
 namespace LionWeb.Core.Test.Utilities.Comparer;
 
 using Core.Utilities;
+using Languages.Generated.V2024_1.Shapes.M2;
 using M2;
 
 [TestClass]
@@ -70,6 +71,43 @@ public class ComparerNodeTests : ComparerTestsBase
             parent,
             new IncompatibleClassifierDifference(left, right) { Parent = parent }
         );
+    }
+
+    [TestMethod]
+    public void Annotation_Concept_Ignored_Same()
+    {
+        var left = lF.NewDocumentation("b");
+        var right = new LenientNode("r", ShapesLanguage.Instance.Documentation);
+
+        Comparer comparer = new Comparer([(IReadableNode?)left], [right])
+        {
+            BehaviorConfig = new()
+            {
+                CompareCompatibleClassifier = false
+            }
+        };
+        Assert.IsTrue(comparer.AreEqual(), comparer.ToMessage(OutputConfig));
+    }
+
+    [TestMethod]
+    public void Annotation_Concept_Ignored_Different()
+    {
+        var left = lF.NewDocumentation("b");
+        var right = rF.NewLine("a");
+
+        Comparer comparer = new Comparer([(IReadableNode?)left], [right])
+        {
+            BehaviorConfig = new()
+            {
+                CompareCompatibleClassifier = false
+            }
+        };
+        
+        var actual = comparer.Compare().Distinct().ToList();
+        var parent = new NodeDifference(left, right);
+        List<IDifference> expected = [parent, new ClassifierDifference(left, right) { Parent = parent } ];
+
+        CollectionAssert.AreEqual(expected, actual, comparer.ToMessage(OutputConfig));
     }
 
     [TestMethod]