+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres _loosely_ to [Semantic Versioning](https://semver.org/s
 * Introduced `SerializerBuilder`.
 * Introduced `LanguageDeserializerBuilder`.
 * Introduced delegating implementations for `SerializerHandler`, `DeserializerHandler`, `LanguageDeserializerHandler`.
+* Added async variant `JsonUtils.WriteNodesToStreamAsync()`.
+* Introduced `ComparerBehaviorConfig.CompareCompatibleClassifier` to configure whether 
+  instances of C# `Annotation` and `Concept` should be considered different.
 * Via `ReflectiveBaseNodeFactory.CreateEnumLiteral()` dynamically created C# enums 
   now have proper `LionCoreMetaPointer` attributes.
 ### Fixed